Check preview feature when looking up related files
