+++ packages/schema-files-loader/helpers/build.ts
@@ -5,7 +5,7 @@ void build([
     name: 'default',
     entryPoints: ['src/index.ts'],
     outfile: 'dist/index',
-    external: ['fs-extra'],
+    external: ['fs-extra', '@prisma/prisma-schema-wasm'],
     bundle: true,
     emitTypes: true,
   },
