+++ packages/schema-files-loader/package.json
@@ -22,6 +22,7 @@
   ],
   "sideEffects": false,
   "dependencies": {
+    "@prisma/prisma-schema-wasm": "5.14.0-17.56ca112d5a19c9925b53af75c3c6b7ada97f9f85",
     "fs-extra": "11.1.1"
   },
   "devDependencies": {
