+++ packages/schema-files-loader/src/__fixtures__/related-feature/a.prisma
@@ -1,0 +1,4 @@
+generator client {
+    provider        = "prisma-client-js"
+    previewFeatures = ["prismaSchemaFolder"]
+}
