+++ packages/schema-files-loader/src/__fixtures__/related-feature/b.prisma
@@ -1,0 +1,1 @@
+// this is b
