+++ packages/schema-files-loader/src/__fixtures__/related-no-feature/a.prisma
@@ -1,0 +1,3 @@
+generator client {
+    provider = "prisma-client-js"
+}
