+++ packages/schema-files-loader/src/index.ts
@@ -1,2 +1,4 @@
+export * from './loadRelatedSchemaFiles'
+export * from './loadSchemaFiles'
 export * from './resolver'
-export * from './schema-files-loader'
+export * from './usesPrismaSchemaFolder'
