+++ packages/schema-files-loader/src/loadSchemaFiles.ts
@@ -2,7 +2,7 @@ import path from 'node:path'
 
 import { FilesResolver, realFsResolver } from './resolver'
 
-type LoadedFile = [filePath: string, content: string]
+export type LoadedFile = [filePath: string, content: string]
 
 type InternalValidatedEntry =
   | {
@@ -55,34 +55,4 @@ async function validateFilePath(fullPath: string, filesResolver: FilesResolver):
     return validateFilePath(realPath, filesResolver)
   }
   return { valid: false }
-}
-
-/**
- * Given a single file path, returns
- * all files composing the same schema
- * @param filePath
- * @param filesResolver
- * @returns
- */
-export function loadRelatedSchemaFiles(
-  filePath: string,
-  filesResolver: FilesResolver = realFsResolver,
-): Promise<LoadedFile[]> {
-  // TODO: this should read preview features from the found files.
-  // If `prismaSchemaFolder` is not enabled, it should only return the entry for `filePaths`
-  // and not any other file
-  return loadSchemaFiles(path.dirname(filePath), filesResolver)
-}
-
-// Minimal subset of the return type of `getConfig` from `@prisma/internals`, used for `usesPrismaSchemaFolder`.
-// This is a simplified version of the actual `ConfigMetaFormat` type, which isn't imported to avoid circular dependencies.
-type ConfigMetaFormat = {
-  generators: Array<{
-    previewFeatures: string[]
-  }>
-}
-
-export function usesPrismaSchemaFolder(config: ConfigMetaFormat): boolean {
-  const previewFeatures = config.generators.find((g) => g.previewFeatures.length > 0)?.previewFeatures
-  return (previewFeatures || []).includes('prismaSchemaFolder')
 }
