+++ packages/schema-files-loader/src/usesPrismaSchemaFolder.ts
@@ -1,0 +1,12 @@
+// Minimal subset of the return type of `getConfig` from `@prisma/internals`, used for `usesPrismaSchemaFolder`.
+// This is a simplified version of the actual `ConfigMetaFormat` type, which isn't imported to avoid circular dependencies.
+export type ConfigMetaFormat = {
+  generators: Array<{
+    previewFeatures: string[]
+  }>
+}
+
+export function usesPrismaSchemaFolder(config: ConfigMetaFormat): boolean {
+  const previewFeatures = config.generators.find((g) => g.previewFeatures.length > 0)?.previewFeatures
+  return (previewFeatures || []).includes('prismaSchemaFolder')
+}
